requirements.txt
tensorflow>=2.19,<2.20
numpy>=1.26
matplotlib>=3.8
Pillow>=10.0
pandas>=2.2
